+++ Engine/shader/texVertex.vert
@@ -11,7 +11,7 @@ layout(location = 3) in float isTex;
 
 layout(location = 0) out vec2 o_uv;
 layout(location = 1) out vec4 o_col;
-layout(location = 2) out float outTexIndex;
+layout(location = 2) out int outTexIndex;
 layout(location = 3) out float outIsTex;
 
 struct Matrix
@@ -19,7 +19,7 @@ struct Matrix
     mat4 model;
     mat4 view;
     mat4 projection;
-    float texIndex;
+    int texIndex;
 };
 
 layout(set = 0, binding = 0) uniform uniformMatrix
